+++ src/Generator/ReferencedType/ReferencedTypeClass.php
@@ -61,11 +61,21 @@ readonly class ReferencedTypeClass implements ReferencedTypeInterface
         return "{$expr} instanceof {$cls}";
     }
 
+    public function typeAssertionConditions(string $expr): array
+    {
+        return [$this->typeAssertionExpr($expr)];
+    }
+
     public function inputAssertionExpr(string $expr): string
     {
         $cls = $this->relativeName();
         $VALIDATE_INPUT = MethodNames::VALIDATE_INPUT;
         return "{$cls}::{$VALIDATE_INPUT}({$expr}, true)";
+    }
+
+    public function inputAssertionConditions(string $expr): array
+    {
+        return [$this->inputAssertionExpr($expr)];
     }
 
     public function inputMappingExpr(string $expr, bool $asserted = false): string